% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawrence_books.R
\name{lawrence_books}
\alias{lawrence_books}
\title{Tidy data frame of 6 of D. H. Lawrence's novels}
\usage{
lawrence_books()
}
\value{
A data frame with two columns: \code{text} and \code{book}
}
\description{
Returns a tidy data frame of 6 of D. H. Lawrence's completed, published novels with
two columns: \code{text}, which contains the text of the novels divided into
elements of up to about 70 characters each, and \code{book}, which contains the titles of
the novels as a factor in order of publication.
}
\examples{

library(dplyr)

lawrence_books() \%>\% group_by(book) \%>\%
     summarise(total_lines = n())

}
