% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{lostgirl}
\alias{lostgirl}
\title{The text of D. H. Lawrence's novel "The Lost Girl"}
\format{
A character vector with 15742 elements
}
\source{
\url{http://www.gutenberg.org/ebooks/23727}
}
\usage{
lostgirl
}
\description{
A dataset containing the text of D. H. Lawrence's 1920 novel "The Lost
Girl". The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
