% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{peacock}
\alias{peacock}
\title{The text of D. H. Lawrence's novel "The White Peacock"}
\format{A character vector with 14626 elements}
\source{
\url{http://www.gutenberg.org/ebooks/38561}
}
\usage{
peacock
}
\description{
A dataset containing the text of D. H. Lawrence's 1911 novel "The White
Peacock". The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}

