% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{rainbow}
\alias{rainbow}
\title{The text of D. H. Lawrence's novel "The Rainbow"}
\format{
A character vector with 20962 elements
}
\source{
\url{http://www.gutenberg.org/ebooks/28948}
}
\usage{
rainbow
}
\description{
A dataset containing the text of D. H. Lawrence's 1915 novel "The Rainbow".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each. 
(Some elements are blank.)
}
\keyword{datasets}
