% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{womeninlove}
\alias{womeninlove}
\title{The text of D. H. Lawrence's novel "Women in Love"}
\format{A character vector with 20157 elements}
\source{
\url{http://www.gutenberg.org/ebooks/4240}
}
\usage{
womeninlove
}
\description{
A dataset containing the text of D. H. Lawrence's novel "Women in Love",
published in 1920.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}

